Added authentication

+++ BlogService/src/main/java/nl/fontys/microserviceblog/controller/BlogController.java
@@ -1,8 +1,12 @@
 package nl.fontys.microserviceblog.controller;
 
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.JwtParser;
+import io.jsonwebtoken.Jwts;
 import nl.fontys.microserviceblog.dao.BlogEntryRepository;
 import nl.fontys.microserviceblog.model.BlogEntry;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -37,17 +41,41 @@ public class BlogController {
     }
 
     @PostMapping()
-    private BlogEntry createBlogEntry(@RequestBody BlogEntry entry) {
-        return this.blogEntryRepository.save(entry);
+    private ResponseEntity createBlogEntry(@RequestBody BlogEntry entry, @RequestHeader("Authentication") String bearer) {
+        String token = bearer.replace("Bearer ", "");
+        Claims claims = Jwts.parser().setSigningKey("secret").parseClaimsJws(token).getBody();
+        int id = Integer.parseInt(claims.get("authorId").toString());
+
+        if (id != entry.getAuthorId()) {
+            return ResponseEntity.unprocessableEntity().build();
+        }
+
+        return ResponseEntity.ok(this.blogEntryRepository.save(entry));
     }
 
     @PutMapping
-    private BlogEntry updateBlogEntry(@RequestBody BlogEntry entry) {
-        return this.blogEntryRepository.save(entry);
+    private ResponseEntity updateBlogEntry(@RequestBody BlogEntry entry, @RequestHeader("Authentication") String bearer) {
+        String token = bearer.replace("Bearer ", "");
+        Claims claims = Jwts.parser().setSigningKey("secret").parseClaimsJws(token).getBody();
+        int id = Integer.parseInt(claims.get("authorId").toString());
+
+        if (id != entry.getAuthorId()) {
+            return ResponseEntity.unprocessableEntity().build();
+        }
+
+        return ResponseEntity.ok(this.blogEntryRepository.save(entry));
     }
 
     @DeleteMapping
-    private ResponseEntity deleteBlogEntry(@RequestBody BlogEntry entry) {
+    private ResponseEntity deleteBlogEntry(@RequestBody BlogEntry entry, @RequestHeader("Authentication") String bearer) {
+        String token = bearer.replace("Bearer ", "");
+        Claims claims = Jwts.parser().setSigningKey("secret").parseClaimsJws(token).getBody();
+        int id = Integer.parseInt(claims.get("authorId").toString());
+
+        if (id != entry.getAuthorId()) {
+            return ResponseEntity.unprocessableEntity().build();
+        }
+
         blogEntryRepository.delete(entry);
         return ResponseEntity.ok().build();
     }
